+++ src/client.ts
@@ -10,6 +10,7 @@ export interface Client<E extends EmissionMap, D extends z.SomeZodObject>
   /** @alias Socket.id */
   id: Socket["id"];
   handshake: Socket["handshake"];
+  getRequest: () => Socket["request"];
   /** @desc Returns the list of the rooms the client in */
   getRooms: () => string[];
   /**